* Removed `async` and `sync` features * Refactored endpoints mapping now it lives in `Router` trait * Refactored middleware mapping, now it lives in `Middleware` trait
