+++ examples/headers.rs
@@ -1,9 +1,9 @@
 ﻿use volga::{
     App,
+    Router,
     ok,
     headers,
     Headers,
-    EndpointsMapping,
     Results,
     ResponseContext
 };
