+++ examples/middleware.rs
@@ -1,8 +1,8 @@
 ﻿use volga::{
     App, 
+    Router, 
+    Middleware,
     ok, 
-    EndpointsMapping, 
-    AsyncMiddlewareMapping
 };
 
 #[tokio::main]
