+++ src/app/endpoints/handlers.rs
@@ -1,5 +1,7 @@
-﻿use futures_util::future::BoxFuture;
-use std::{sync::Arc, future::Future};
+﻿use std::{sync::Arc, future::Future};
+
+use futures_util::future::BoxFuture;
+
 use crate::{HttpResult, HttpRequest};
 use crate::app::endpoints::args::FromRequest;
 
@@ -7,36 +9,6 @@ pub(crate) type RouteHandler = Arc<dyn Handler + Send + Sync>;
 
 pub(crate) trait Handler {
     fn call(&self, req: HttpRequest) -> BoxFuture<HttpResult>;
-}
-
-#[cfg(feature = "async")]
-pub(crate) struct AsyncHandler<F>(pub F);
-
-#[cfg(feature = "async")]
-impl<F, Fut> Handler for AsyncHandler<F>
-where
-    F: Fn(HttpRequest) -> Fut,
-    Fut: Future<Output = HttpResult> + Send + 'static,
-{
-    #[inline]
-    fn call(&self, req: HttpRequest) -> BoxFuture<HttpResult> {
-        Box::pin(self.0(req))
-    }
-}
-
-#[cfg(feature = "sync")]
-pub(crate) struct SyncHandler<F>(pub F);
-
-#[cfg(feature = "sync")]
-impl<F> Handler for SyncHandler<F>
-where
-    F: Fn(HttpRequest) -> HttpResult,
-{
-    #[inline]
-    fn call(&self, req: HttpRequest) -> BoxFuture<HttpResult> {
-        let response = self.0(req);
-        Box::pin(async move { response })
-    }
 }
 
 pub(crate) struct Func<F, Args>
