--- src/app/endpoints/mapping.rs
@@ -1,4 +1,0 @@
-﻿#[cfg(feature = "async")]
-pub mod asynchronous;
-#[cfg(feature = "sync")]
-pub mod synchronous;
