+++ src/lib.rs
@@ -40,7 +40,7 @@ pub mod app;
 #[cfg(test)]
 pub mod test_utils;
 
-pub use crate::app::App;
+pub use crate::app::{App, router::Router};
 pub use crate::app::results::{HttpResponse, HttpResult, HttpHeaders, Results, ResponseContext};
 pub use crate::app::request::HttpRequest;
 
@@ -48,13 +48,8 @@ pub use crate::app::request::HttpRequest;
 pub use crate::app::http_context::HttpContext;
 
 #[cfg(feature = "middleware")]
-pub use crate::app::middlewares::{Next, mapping::asynchronous::AsyncMiddlewareMapping};
+pub use crate::app::middlewares::{Next, Middleware};
 
-#[cfg(feature = "async")]
-pub use crate::app::endpoints::mapping::asynchronous::AsyncEndpointsMapping;
-#[cfg(feature = "async")]
-pub use crate::app::endpoints::mapping::asynchronous::EndpointsMapping;
-#[cfg(feature = "async")]
 pub use crate::app::endpoints::args::{
     path::Path,
     query::Query,
@@ -63,8 +58,5 @@ pub use crate::app::endpoints::args::{
     file::File,
     cancellation_token::CancellationToken,
 };
-
-#[cfg(feature = "sync")]
-pub use crate::app::endpoints::mapping::synchronous::SyncEndpointsMapping;
 
 pub use crate::app::body::BoxBody;
