--- tests/app_sync_mapping_tests.rs
@@ -1,117 +1,0 @@
-﻿use volga::{App, Results};
-use volga::SyncEndpointsMapping;
-
-#[tokio::test]
-async fn it_maps_to_get_request() {
-    tokio::spawn(async {
-        let mut app = App::new().bind("127.0.0.1:7879");
-        app.map_get("/test", |_req| {
-            Results::text("Pass!")
-        });
-       app.run().await
-    });
-
-    let response = tokio::spawn(async {
-        let client = if cfg!(all(feature = "http1", not(feature = "http2"))) {
-            reqwest::Client::builder().http1_only().build().unwrap()
-        } else {
-            reqwest::Client::builder().http2_prior_knowledge().build().unwrap()
-        };
-        client.get("http://127.0.0.1:7879/test").send().await
-    }).await.unwrap().unwrap();
-
-    assert!(response.status().is_success());
-    assert_eq!(response.text().await.unwrap(), "Pass!");
-}
-
-#[tokio::test]
-async fn it_maps_to_post_request() {
-    tokio::spawn(async {
-        let mut app = App::new().bind("127.0.0.1:7880");
-        app.map_post("/test", |_req| {
-            Results::text("Pass!")
-        });
-       app.run().await
-    });
-
-    let response = tokio::spawn(async {
-        let client = if cfg!(all(feature = "http1", not(feature = "http2"))) {
-            reqwest::Client::builder().http1_only().build().unwrap()
-        } else {
-            reqwest::Client::builder().http2_prior_knowledge().build().unwrap()
-        };
-        client.post("http://127.0.0.1:7880/test").send().await
-    }).await.unwrap().unwrap();
-
-    assert!(response.status().is_success());
-    assert_eq!(response.text().await.unwrap(), "Pass!");
-}
-
-#[tokio::test]
-async fn it_maps_to_put_request() {
-    tokio::spawn(async {
-        let mut app = App::new().bind("127.0.0.1:7881");
-        app.map_put("/test", |_req| {
-            Results::text("Pass!")
-        });
-       app.run().await
-    });
-
-    let response = tokio::spawn(async {
-        let client = if cfg!(all(feature = "http1", not(feature = "http2"))) {
-            reqwest::Client::builder().http1_only().build().unwrap()
-        } else {
-            reqwest::Client::builder().http2_prior_knowledge().build().unwrap()
-        };
-        client.put("http://127.0.0.1:7881/test").send().await
-    }).await.unwrap().unwrap();
-
-    assert!(response.status().is_success());
-    assert_eq!(response.text().await.unwrap(), "Pass!");
-}
-
-#[tokio::test]
-async fn it_maps_to_patch_request() {
-    tokio::spawn(async {
-        let mut app = App::new().bind("127.0.0.1:7882");
-        app.map_patch("/test", |_req| {
-            Results::text("Pass!")
-        });
-       app.run().await
-    });
-
-    let response = tokio::spawn(async {
-        let client = if cfg!(all(feature = "http1", not(feature = "http2"))) {
-            reqwest::Client::builder().http1_only().build().unwrap()
-        } else {
-            reqwest::Client::builder().http2_prior_knowledge().build().unwrap()
-        };
-        client.patch("http://127.0.0.1:7882/test").send().await
-    }).await.unwrap().unwrap();
-
-    assert!(response.status().is_success());
-    assert_eq!(response.text().await.unwrap(), "Pass!");
-}
-
-#[tokio::test]
-async fn it_maps_to_delete_request() {
-    tokio::spawn(async {
-        let mut app = App::new().bind("127.0.0.1:7883");
-        app.map_delete("/test", |_req| {
-            Results::text("Pass!")
-        });
-       app.run().await
-    });
-
-    let response = tokio::spawn(async {
-        let client = if cfg!(all(feature = "http1", not(feature = "http2"))) {
-            reqwest::Client::builder().http1_only().build().unwrap()
-        } else {
-            reqwest::Client::builder().http2_prior_knowledge().build().unwrap()
-        };
-        client.delete("http://127.0.0.1:7883/test").send().await
-    }).await.unwrap().unwrap();
-
-    assert!(response.status().is_success());
-    assert_eq!(response.text().await.unwrap(), "Pass!");
-}
